Get rid of x64 compile warnings

+++ TGA.h
@@ -9,7 +9,7 @@ class TGA {
 public:
   TGA() : data_(NULL) {}
   ~TGA() { delete[] data_; }
-  TGA(size_t width, size_t height) :
+  TGA(int width, int height) :
       width_(width), height_(height),
       data_(new unsigned char[width*height*3]) {}
   bool readFile(const char* filename) {
@@ -28,8 +28,8 @@ public:
                 " : unsupported TGA format, only 24bpp supported\n");
         break;
       }
-      size_t width = header[13] * 256 + header[12];
-      size_t height = header[15] * 256 + header[14];
+      int width = header[13] * 256 + header[12];
+      int height = header[15] * 256 + header[14];
       if (width > 32768 || height > 32768) {
         fprintf(stderr, __FUNCTION__
                 " : oversized TGA image not supported\n");
@@ -71,7 +71,7 @@ public:
     return result;
   }
 #ifdef GL_BGR
-  bool readFramebuffer(size_t width, size_t height, int viewportOffset[2]) {
+  bool readFramebuffer(int width, int height, int viewportOffset[2]) {
     width_ = width;
     height_ = height;
     delete[] data_;
@@ -92,13 +92,13 @@ public:
     data_[x*3+width_*3*y+2] = GLSL::clamp(256*col.x, 0, 255);  // R
   }
 #endif
-  size_t width() { return width_; }
-  size_t height() { return height_; }
+  int width() { return width_; }
+  int height() { return height_; }
   unsigned char* data() { return data_; }
 
 private:
-  size_t width_;
-  size_t height_;
+  int width_;
+  int height_;
   unsigned char* data_;
 };
